+++ src/main/java/com/example/jomajomadelivery/account/oauth2/service/handler/CustomAuthenticationSuccessHandler.java
@@ -1,8 +1,8 @@
-package com.example.jomajomadelivery.account.auth.oauth2.handler;
+package com.example.jomajomadelivery.account.oauth2.service.handler;
 
-import com.example.jomajomadelivery.account.auth.jwt.TokenProvider;
-import com.example.jomajomadelivery.account.auth.oauth2.CustomOAuth2User;
-import com.example.jomajomadelivery.account.auth.oauth2.SocialProvider;
+import com.example.jomajomadelivery.account.jwt.TokenProvider;
+import com.example.jomajomadelivery.account.oauth2.service.CustomOAuth2User;
+import com.example.jomajomadelivery.account.oauth2.service.SocialProvider;
 import com.example.jomajomadelivery.account.auth.service.UserAuthService;
 import com.example.jomajomadelivery.user.entity.User;
 import jakarta.servlet.ServletException;
